remove leaderboard table
Previously rankings were generated once every 60 seconds and the results
stored in a table.   Now the most recent ranking will be stored in the
submission table and the history will be stored in the game_player
table.  Rank movements for a bot due to other bots changing rank will
not be captured.  The change in ranks between games due to the results
of other games will be captured as a whole in the difference between
the last rank_after and the next rank_before in the submission table.