speed up of matchmaking
The exclusion list is now making sure there are no repeat game for all
players, not just the seed player, for the last game.  The percentage
of users logic is not needed since the list is larger.

Positions are placed randomly because we now have symmetric maps and
they don't really matter.

Games played and a random number are used as fallback order values for
picking the opponents.